chore: Move sign and send subtask to a flow
