+++ packages/devtools/src/index.ts
@@ -1,5 +1,6 @@
 export * from './common'
 export * from './docker'
+export * from './flows'
 export * from './omnigraph'
 export * from './transactions'
 export * from './types'
